server/pyproject.toml
[project]
name = "teal-care-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
]
